+++ otelconf/x/log_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package otelconf
+package x
 
 import (
 	"bytes"
@@ -258,7 +258,7 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								CaFile: ptr(filepath.Join("testdata", "ca.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "ca.crt")),
 							},
 						},
 					},
@@ -276,7 +276,7 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								CaFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -310,8 +310,8 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								KeyFile:  ptr(filepath.Join("testdata", "bad_cert.crt")),
-								CertFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								KeyFile:  ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
+								CertFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -417,7 +417,7 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								CaFile: ptr(filepath.Join("testdata", "ca.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "ca.crt")),
 							},
 						},
 					},
@@ -435,7 +435,7 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								CaFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -453,8 +453,8 @@ func TestLogProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								KeyFile:  ptr(filepath.Join("testdata", "bad_cert.crt")),
-								CertFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								KeyFile:  ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
+								CertFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -633,7 +633,9 @@ func TestLogProcessor(t *testing.T) {
 			name: "simple/otlp_file",
 			processor: LogRecordProcessor{
 				Simple: &SimpleLogRecordProcessor{
-					Exporter: LogRecordExporter{},
+					Exporter: LogRecordExporter{
+						OTLPFileDevelopment: &ExperimentalOTLPFileExporter{},
+					},
 				},
 			},
 			wantErrT: newErrInvalid("otlp_file/development"),
@@ -735,9 +737,9 @@ func TestLoggerProviderOptions(t *testing.T) {
 }
 
 func Test_otlpGRPCLogExporter(t *testing.T) {
-	if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
-		// TODO (#8115): Fix the flakiness on Windows and MacOS.
-		t.Skip("Test is flaky on Windows and MacOS.")
+	if runtime.GOOS == "windows" {
+		// TODO (#7446): Fix the flakiness on Windows.
+		t.Skip("Test is flaky on Windows.")
 	}
 	type args struct {
 		ctx        context.Context
@@ -775,7 +777,7 @@ func Test_otlpGRPCLogExporter(t *testing.T) {
 					Compression: ptr("gzip"),
 					Timeout:     ptr(5000),
 					Tls: &GrpcTls{
-						CaFile: ptr("testdata/server-certs/server.crt"),
+						CaFile: ptr("../testdata/server-certs/server.crt"),
 					},
 					Headers: []NameStringValuePair{
 						{Name: "test", Value: ptr("test1")},
@@ -784,7 +786,7 @@ func Test_otlpGRPCLogExporter(t *testing.T) {
 			},
 			grpcServerOpts: func() ([]grpc.ServerOption, error) {
 				opts := []grpc.ServerOption{}
-				tlsCreds, err := credentials.NewServerTLSFromFile("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				tlsCreds, err := credentials.NewServerTLSFromFile("../testdata/server-certs/server.crt", "../testdata/server-certs/server.key")
 				if err != nil {
 					return nil, err
 				}
@@ -800,9 +802,9 @@ func Test_otlpGRPCLogExporter(t *testing.T) {
 					Compression: ptr("gzip"),
 					Timeout:     ptr(5000),
 					Tls: &GrpcTls{
-						CaFile:   ptr("testdata/server-certs/server.crt"),
-						KeyFile:  ptr("testdata/client-certs/client.key"),
-						CertFile: ptr("testdata/client-certs/client.crt"),
+						CaFile:   ptr("../testdata/server-certs/server.crt"),
+						KeyFile:  ptr("../testdata/client-certs/client.key"),
+						CertFile: ptr("../testdata/client-certs/client.crt"),
 					},
 					Headers: []NameStringValuePair{
 						{Name: "test", Value: ptr("test1")},
@@ -811,11 +813,11 @@ func Test_otlpGRPCLogExporter(t *testing.T) {
 			},
 			grpcServerOpts: func() ([]grpc.ServerOption, error) {
 				opts := []grpc.ServerOption{}
-				cert, err := tls.LoadX509KeyPair("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				cert, err := tls.LoadX509KeyPair("../testdata/server-certs/server.crt", "../testdata/server-certs/server.key")
 				if err != nil {
 					return nil, err
 				}
-				caCert, err := os.ReadFile("testdata/ca.crt")
+				caCert, err := os.ReadFile("../testdata/ca.crt")
 				if err != nil {
 					return nil, err
 				}