+++ otelconf/x/resource.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package otelconf // import "go.opentelemetry.io/contrib/otelconf"
+package x // import "go.opentelemetry.io/contrib/otelconf/x"
 
 import (
 	"context"
@@ -11,6 +11,24 @@ import (
 
 	"go.opentelemetry.io/contrib/otelconf/internal/kv"
 )
+
+func resourceOpts(detectors []ExperimentalResourceDetector) []resource.Option {
+	opts := []resource.Option{}
+	for _, d := range detectors {
+		if d.Container != nil {
+			opts = append(opts, resource.WithContainer())
+		}
+		if d.Host != nil {
+			opts = append(opts, resource.WithHost(), resource.WithHostID())
+		}
+		if d.Process != nil {
+			opts = append(opts, resource.WithProcess())
+		}
+		// TODO: implement service:
+		// Waiting on https://github.com/open-telemetry/opentelemetry-go/pull/7642
+	}
+	return opts
+}
 
 func newResource(r *Resource) (*resource.Resource, error) {
 	if r == nil {
@@ -29,6 +47,10 @@ func newResource(r *Resource) (*resource.Resource, error) {
 	opts := []resource.Option{
 		resource.WithAttributes(attrs...),
 		resource.WithSchemaURL(schema),
+	}
+
+	if r.DetectionDevelopment != nil {
+		opts = append(opts, resourceOpts(r.DetectionDevelopment.Detectors)...)
 	}
 
 	return resource.New(context.Background(), opts...)