+++ otelconf/x/trace_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package otelconf
+package x
 
 import (
 	"bytes"
@@ -15,7 +15,6 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
-	"runtime"
 	"testing"
 	"time"
 
@@ -360,7 +359,7 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								CaFile: ptr(filepath.Join("testdata", "ca.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "ca.crt")),
 							},
 						},
 					},
@@ -378,7 +377,7 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								CaFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -396,8 +395,8 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &GrpcTls{
-								KeyFile:  ptr(filepath.Join("testdata", "bad_cert.crt")),
-								CertFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								KeyFile:  ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
+								CertFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -519,7 +518,7 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								CaFile: ptr(filepath.Join("testdata", "ca.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "ca.crt")),
 							},
 						},
 					},
@@ -537,7 +536,7 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								CaFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								CaFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -555,8 +554,8 @@ func TestSpanProcessor(t *testing.T) {
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
 							Tls: &HttpTls{
-								KeyFile:  ptr(filepath.Join("testdata", "bad_cert.crt")),
-								CertFile: ptr(filepath.Join("testdata", "bad_cert.crt")),
+								KeyFile:  ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
+								CertFile: ptr(filepath.Join("..", "testdata", "bad_cert.crt")),
 							},
 						},
 					},
@@ -735,7 +734,9 @@ func TestSpanProcessor(t *testing.T) {
 			name: "simple/otlp_file",
 			processor: SpanProcessor{
 				Simple: &SimpleSpanProcessor{
-					Exporter: SpanExporter{},
+					Exporter: SpanExporter{
+						OTLPFileDevelopment: &ExperimentalOTLPFileExporter{},
+					},
 				},
 			},
 			wantErrT: newErrInvalid("otlp_file/development"),
@@ -899,10 +900,6 @@ func TestSampler(t *testing.T) {
 }
 
 func Test_otlpGRPCTraceExporter(t *testing.T) {
-	if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
-		// TODO (#8115): Fix the flakiness on Windows and MacOS.
-		t.Skip("Test is flaky on Windows and MacOS.")
-	}
 	type args struct {
 		ctx        context.Context
 		otlpConfig *OTLPGrpcExporter
@@ -939,7 +936,7 @@ func Test_otlpGRPCTraceExporter(t *testing.T) {
 					Compression: ptr("gzip"),
 					Timeout:     ptr(5000),
 					Tls: &GrpcTls{
-						CaFile: ptr("testdata/server-certs/server.crt"),
+						CaFile: ptr("../testdata/server-certs/server.crt"),
 					},
 					Headers: []NameStringValuePair{
 						{Name: "test", Value: ptr("test1")},
@@ -948,7 +945,7 @@ func Test_otlpGRPCTraceExporter(t *testing.T) {
 			},
 			grpcServerOpts: func() ([]grpc.ServerOption, error) {
 				opts := []grpc.ServerOption{}
-				tlsCreds, err := credentials.NewServerTLSFromFile("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				tlsCreds, err := credentials.NewServerTLSFromFile("../testdata/server-certs/server.crt", "../testdata/server-certs/server.key")
 				if err != nil {
 					return nil, err
 				}
@@ -964,9 +961,9 @@ func Test_otlpGRPCTraceExporter(t *testing.T) {
 					Compression: ptr("gzip"),
 					Timeout:     ptr(5000),
 					Tls: &GrpcTls{
-						CaFile:   ptr("testdata/server-certs/server.crt"),
-						KeyFile:  ptr("testdata/client-certs/client.key"),
-						CertFile: ptr("testdata/client-certs/client.crt"),
+						CaFile:   ptr("../testdata/server-certs/server.crt"),
+						KeyFile:  ptr("../testdata/client-certs/client.key"),
+						CertFile: ptr("../testdata/client-certs/client.crt"),
 					},
 					Headers: []NameStringValuePair{
 						{Name: "test", Value: ptr("test1")},
@@ -975,11 +972,11 @@ func Test_otlpGRPCTraceExporter(t *testing.T) {
 			},
 			grpcServerOpts: func() ([]grpc.ServerOption, error) {
 				opts := []grpc.ServerOption{}
-				cert, err := tls.LoadX509KeyPair("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				cert, err := tls.LoadX509KeyPair("../testdata/server-certs/server.crt", "../testdata/server-certs/server.key")
 				if err != nil {
 					return nil, err
 				}
-				caCert, err := os.ReadFile("testdata/ca.crt")
+				caCert, err := os.ReadFile("../testdata/ca.crt")
 				if err != nil {
 					return nil, err
 				}